+++ API/authchecks.cjs
