+++ main.cjs
@@ -9,7 +9,7 @@ const {hashPassword,comparePassword,compareDBbusiness,compareDBbusinessv2,compar
 const {tokenChecker,TokenGen,TokenGenEnt,TokenGenVend, TokenGenAdmin, st} = require ("./tokenchecker.cjs");
 require('dotenv').config({ path: 'process.env' });
 const {LocalStorage} = require('node-localstorage');
-const authcheck = require('./API/authcheck.cjs');
+const authcheck = require('./API/authchecks.cjs');
 
 const Prodotto = require('./classes/prodotto.cjs');
 
@@ -29,18 +29,18 @@ const DBFormVend = require ('./models/upgradeModel.cjs');
 const DBAdmin = require('./models/adminModel.cjs');
 
 //API imports
-const carrelli = require('./API/carrello.cjs');
-const promozioni = require('./API/promozione.cjs');
-const venditori = require('./API/venditore.cjs');
-const prodotti = require('./API/prodotto.cjs');
-const accounts = require('./API/account.cjs');
+const carrelli = require('./API/carrelli.cjs');
+const promozioni = require('./API/promozioni.cjs');
+const venditori = require('./API/venditori.cjs');
+const prodotti = require('./API/prodotti.cjs');
+const accounts = require('./API/accounts.cjs');
 const tags = require('./API/tags.cjs');
 const citytags = require('./API/citytags.cjs');
-const order = require('./API/order.cjs');
-const mail = require('./API/mail.cjs');
+const order = require('./API/orders.cjs');
+const mail = require('./API/mails.cjs');
 const clienti = require('./API/clienti.cjs');
-const upgrade = require('./API/upgrade.cjs');
-const imprenditore = require('./API/imprenditore.cjs');
+const upgrade = require('./API/upgrades.cjs');
+const imprenditore = require('./API/imprenditori.cjs');
 const consegne = require('./API/consegne.cjs')
 const app = express();
 const port = process.env.PORT || 3000;
