+++ public/consegne-admin.html
@@ -69,6 +69,7 @@
         }
         .approve-btn {
             margin-top: 8px;
+            align-items: center;
             background: #218838;
             color: white;
             border: none;
@@ -152,51 +153,69 @@
     async function displayConsegne(consegne) {
         consegneDisplay.innerHTML = '';
         for (const consegna of consegne) {
-            
             if (consegna.status === 'In consegna') {
                 const consegnaDiv = document.createElement('div');
                 consegnaDiv.className = 'ordine';
-                consegnaDiv.innerHTML = `
-                    <strong>Data:</strong> ${new Date(consegna.data).toLocaleDateString()} <br>
-                `;
-                consegna.ordini.forEach(ordine => {
-                    const res = fetch(`/api/v1/orders/${ordine.id}`, {
-                        headers: { 'x-access-token': token }
-                    });
-                    res.then(response => response.json()).then(order => {
-                        consegnaDiv.innerHTML += `
-                            <strong>Venditore:</strong> ${order.venditore} <br>
-                            <strong>Cliente:</strong> ${order.cliente} <br>
+                let inner = `<strong>Data:</strong> ${new Date(consegna.data).toLocaleDateString()} <br>`;
+
+                // Gather all order info
+                for (const ordine of consegna.ordini) {
+                    try {
+                        const orderRes = await fetch(`/api/v1/orders/${ordine.id}`, {
+                            headers: { 'x-access-token': token }
+                        });
+                        const order = await orderRes.json();
+
+                        const vendRes = await fetch(`/api/v1/venditori/${order.venditore}`, {
+                            headers: { 'x-access-token': token }
+                        });
+                        const vend = await vendRes.json();
+
+                        const clientRes = await fetch(`/api/v1/clienti/${order.cliente}`, {
+                            headers: { 'x-access-token': token }
+                        });
+                        const client = await clientRes.json();
+
+                        inner += `
+                            <strong>Venditore:</strong> ${vend.nome + ' ' + vend.cognome} <br>
+                            <strong>Venditore email:</strong> ${vend.email} <br>
+                            <strong>Cliente:</strong> ${client.nome + ' ' + client.cognome} <br>
+                            <strong>Cliente email:</strong> ${client.email} <br>
+                            <strong>Residenza venditore:</strong> ${vend.sede} <br>
                             <strong>Zona:</strong> ${order.zona} <br>
                         `;
-                    }).catch(error => {
-                        console.error('Errore nel recupero dell\'ordine:', error);
-                    });
-                });
-                consegneDisplay.appendChild(consegnaDiv);
+                    } catch (e) {
+                        inner += '<span style="color:red;">Errore nel recupero dei dati dell\'ordine.</span>';
+                        console.error('Errore nel recupero dei dati dell\'ordine:', e);
+                    }
+                }
+
+                consegnaDiv.innerHTML = inner;
 
                 const approveBtn = document.createElement('button');
                 approveBtn.className = 'approve-btn';
                 approveBtn.textContent = 'Approva';
                 approveBtn.onclick = async () => {
                     try {
-                        const response = await fetch(`/api/v1/consegne/${consegna.id}/approva`, {
+                        const response = await fetch(`/api/v1/consegne/${consegna._id}/approva`, {
                             method: 'PUT',
                             headers: {
-                                'Content-Type': 'application/json',
                                 'x-access-token': token
                             }
                         });
                         if (!response.ok) {
+                            alert('Errore nell\'approvazione della consegna.' + response.statusText);
                             throw new Error('Errore nell\'approvazione della consegna');
                         }
                         consegnaDiv.innerHTML += '<span style="color:green;">Consegna approvata!</span>';
+                        window.location.reload();
                     } catch (error) {
                         console.error('Errore nell\'approvazione della consegna:', error);
                         alert('Errore nell\'approvazione della consegna.');
                     }
                 };
                 consegnaDiv.appendChild(approveBtn);
+                consegneDisplay.appendChild(consegnaDiv);
             }
         }
     }
